+++ src/main/kotlin/com/example/voteapp/server/plugins/Routing.kt
@@ -3,7 +3,8 @@ package com.example.voteapp.server.plugins
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import io.ktor.server.application.*
-import com.example.voteapp.server.votings.configureVotingsRouting
+import com.example.voteapp.server.votings.installVotings
+
 
 fun Application.configureRouting() {
 
@@ -16,7 +17,8 @@ fun Application.configureRouting() {
 
 
     // Модули роутинга
-    configureVotingsRouting()
+    installVotings()
+
 }